requirements.txt
# Learnzo backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.2.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5
python-multipart>=0.0.6
requests>=2.31.0
gtts>=2.4.0
python-dotenv>=1.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
